Pause statemachine earlier on snapshot sync
