+++ include/libnuraft/asio_service_options.hxx
@@ -20,6 +20,7 @@ limitations under the License.
 #include <functional>
 #include <string>
 #include <system_error>
+#include <cstdint>
 
 
 typedef struct ssl_ctx_st SSL_CTX;
