+++ include/libnuraft/callback.hxx
@@ -20,6 +20,7 @@ limitations under the License.
 
 #include <functional>
 #include <string>
+#include <cstdint>
 
 namespace nuraft {
